Alterando Identity

+++ src/Identity/Identity.Domain/Services/UserService.cs
@@ -1,4 +1,5 @@
 ﻿using Identity.Domain.Core;
+using Identity.Domain.Interfaces.Identity;
 using Identity.Domain.Interfaces.Repositories;
 using Identity.Domain.Interfaces.Services;
 using Identity.Domain.Models;
@@ -9,12 +10,15 @@ namespace Identity.Domain.Services
     {
         private readonly NotificationContext _notification;
         private readonly IUserRepository _repository;
+        private readonly IUserIdentity _identity;
 
         public UserService(NotificationContext notification,
-            IUserRepository repository)
+            IUserRepository repository,
+            IUserIdentity identity)
         {
             _notification = notification;
             _repository = repository;
+            _identity = identity;
         }       
 
         public async Task<User> GetByIdAsync(string id)
@@ -24,6 +28,22 @@ namespace Identity.Domain.Services
             if (user == null) 
             {
                 _notification.AddNotification("Não foi encontrado nenhum Usuário");
+
+                return null;
+            }
+
+            return user;
+        }
+
+        public async Task<User> GetAsync()
+        {
+            var userId = _identity.GetUserIdFromToken();
+            var user = await _repository.GetByIdAsync(userId);
+
+            if (user == null)
+            {
+                _notification.AddNotification("Não foi encontrado nenhum Usuário");
+
                 return null;
             }
 
@@ -40,8 +60,18 @@ namespace Identity.Domain.Services
             await _repository.UpdateAsync(user);
         }
 
-        public async Task DeleteAsync(User user)
+        public async Task DeleteAsync()
         {
+            var userId = _identity.GetUserIdFromToken();
+            var user = await _repository.GetByIdAsync(userId);
+
+            if (user == null) 
+            {
+                _notification.AddNotification("Não foi possivel encontrar nenhum Usuário");
+
+                return;
+            }
+
             await _repository.DeleteAsync(user);
         }
     }
